+++ RockEngine/src/RockEngine/Renderer/Mesh.cpp
@@ -131,7 +131,7 @@ namespace RockEngine
 				aabb.Min = { FLT_MAX, FLT_MAX, FLT_MAX };
 				aabb.Max = { -FLT_MAX, -FLT_MAX, -FLT_MAX };
 
-				for (size_t i = 0; i < mesh->mNumVertices; i++)
+				for (int i = 0; i < mesh->mNumVertices; i++)
 				{
 					Vertex vertex;
 					vertex.Position = { mesh->mVertices[i].x, mesh->mVertices[i].y, mesh->mVertices[i].z };
