+++ RockEngine/src/RockEngine/Renderer/SceneRenderer.cpp
@@ -193,9 +193,9 @@ namespace RockEngine
 
 		Renderer::Submit([envUnfiltered, envFiltered, cubemapSize]() {
 			const float deltaRoughness = 1.0f / glm::max((float)(envFiltered->GetMipLevelCount() - 1.0f), 1.0f);
-			for (int level = 1, size = cubemapSize / 2; level < envFiltered->GetMipLevelCount(); level++, size /= 2) // <= ?
+			for (u32 level = 1, size = cubemapSize / 2; level < envFiltered->GetMipLevelCount(); level++, size /= 2) // <= ?
 			{
-				const GLuint numGroups = glm::max(1, size / 32);
+				const GLuint numGroups = glm::max(1, (int)size / 32);
 				glBindImageTexture(0, envFiltered->GetRendererID(), level, GL_TRUE, 0, GL_WRITE_ONLY, GL_RGBA16F);
 				glProgramUniform1f(envFilteringShader->GetRendererID(), 0, level * deltaRoughness);
 				glDispatchCompute(numGroups, numGroups, 6);
